chore: Move shared to lib/
